Fix warning in Box.cpp: Add cast to size_t.

+++ src/physics/Box.cpp
@@ -1054,7 +1054,7 @@ void EERIE_PHYSICS_BOX_Create(EERIE_3DOBJ * obj)
 
 	for (size_t k = 0; k < obj->vertexlist.size(); k++)
 	{
-		if (k == obj->origin) continue;
+		if (k == (size_t)obj->origin) continue;
 
 		cubmin.x = min(cubmin.x, obj->vertexlist[k].v.x);
 		cubmin.y = min(cubmin.y, obj->vertexlist[k].v.y);
